Image Encryption Using DES3/Image Encryption Using DES3/requirement.txt
certifi==2020.12.5
chardet==4.0.0
crypto==1.4.1
idna==2.10
Naked==0.1.31
Pillow==8.2.0
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
urllib3==1.26.4
